Make it Right: got the code to compile after moving.
