+++ jb/opencl/generic_reduce_program.hpp
@@ -1,13 +1,13 @@
-#ifndef jb_tde_generic_reduce_program_hpp
-#define jb_tde_generic_reduce_program_hpp
+#ifndef jb_opencl_generic_reduce_program_hpp
+#define jb_opencl_generic_reduce_program_hpp
 
 namespace jb {
-namespace tde {
+namespace opencl {
 
 /// Contains the code for the kernels used in computing the argmax.
 extern char const generic_reduce_program_source[];
 
-} // namespace tde
+} // namespace opencl
 } // namespace jb
 
-#endif // jb_tde_generic_reduce_program_hpp
+#endif // jb_opencl_generic_reduce_program_hpp
